ui: initial draggable() modifier on sample list
